refactor(dashboard): move cross model manager logic to biz
