--- src/dashboard/apigateway/apigateway/apps/programmable_gateway/managers.py
@@ -1,57 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# TencentBlueKing is pleased to support the open source community by making
-# 蓝鲸智云 - API 网关(BlueKing - APIGateway) available.
-# Copyright (C) 2025 Tencent. All rights reserved.
-# Licensed under the MIT License (the "License"); you may not use this file except
-# in compliance with the License. You may obtain a copy of the License at
-#
-#     http://opensource.org/licenses/MIT
-#
-# Unless required by applicable law or agreed to in writing, software distributed under
-# the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
-# either express or implied. See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# We undertake not to change the open source license (MIT license) applicable
-# to the current version of the project delivered to anyone in the future.
-#
-
-from django.db import models
-from django.db.models import Q
-
-
-class ProgrammableGatewayDeployHistoryManager(models.Manager):
-    def filter_deploy_history(
-        self,
-        gateway,
-        query="",
-        stage_id=None,
-        created_by="",
-        time_start=None,
-        time_end=None,
-        order_by=None,
-        fuzzy=False,
-    ):
-        queryset = self.filter(gateway=gateway)
-        # query 不是模型字段，仅支持模糊匹配，如需精确匹配，可使用具体字段
-        if query and fuzzy:
-            queryset = queryset.filter(Q(stage__name__contains=query) | Q(version__contains=query))
-
-        if stage_id:
-            queryset = queryset.filter(stage_id=stage_id)
-
-        if created_by:
-            if fuzzy:
-                queryset = queryset.filter(created_by__contains=created_by)
-            else:
-                queryset = queryset.filter(created_by=created_by)
-
-        if time_start and time_end:
-            # time_start、time_end 须同时存在，否则无效
-            queryset = queryset.filter(created_time__range=(time_start, time_end))
-
-        if order_by:
-            queryset = queryset.order_by(order_by)
-
-        return queryset.distinct()
