+++ src/dashboard/apigateway/apigateway/core/managers.py
@@ -24,7 +24,6 @@ from typing import Any, Dict, List, Optional
 
 from cachetools import TTLCache, cached
 from django.db import models
-from django.db.models import Q
 
 from apigateway.common.constants import CACHE_MAXSIZE, CACHE_TIME_24_HOURS
 from apigateway.core.constants import (
@@ -342,45 +341,6 @@ class ReleasedResourceManager(models.Manager):
             "resource_perm_required": resource_auth_config["resource_perm_required"],
             "user_verified_required": resource_auth_config["auth_verified_required"],
         }
-
-
-class ReleaseHistoryManager(models.Manager):
-    # FIXME: not common, move to views.py
-    def filter_release_history(
-        self,
-        gateway,
-        query="",
-        stage_id=None,
-        created_by="",
-        time_start=None,
-        time_end=None,
-        order_by=None,
-        fuzzy=False,
-    ):
-        queryset = self.filter(gateway=gateway)
-
-        # query 不是模型字段，仅支持模糊匹配，如需精确匹配，可使用具体字段
-        if query and fuzzy:
-            queryset = queryset.filter(Q(stage__name__contains=query) | Q(resource_version__version__contains=query))
-
-        if stage_id:
-            queryset = queryset.filter(stage_id=stage_id)
-
-        if created_by:
-            if fuzzy:
-                queryset = queryset.filter(created_by__contains=created_by)
-            else:
-                queryset = queryset.filter(created_by=created_by)
-
-        if time_start and time_end:
-            # time_start、time_end 须同时存在，否则无效
-            queryset = queryset.filter(created_time__range=(time_start, time_end))
-
-        if order_by:
-            queryset = queryset.order_by(order_by)
-
-        # Select related data_plane for ReleaseHistory to avoid N+1 queries
-        return queryset.select_related("data_plane").distinct()
 
 
 class PublishEventManager(models.Manager):
