+++ src/dashboard/apigateway/apigateway/tests/biz/release/test_release.py
@@ -15,8 +15,10 @@
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
 #
+import datetime
 from unittest.mock import patch
 
+import pytest
 from ddf import G
 
 from apigateway.biz.release import ReleaseHandler
@@ -27,11 +29,92 @@ from apigateway.core.constants import (
     ReleaseHistoryStatusEnum,
     StageStatusEnum,
 )
-from apigateway.core.models import PublishEvent, Release, Stage
+from apigateway.core.models import Gateway, PublishEvent, Release, ReleaseHistory, ResourceVersion, Stage
+from apigateway.tests.utils.testing import dummy_time
 from apigateway.utils.time import now_datetime
+
+pytestmark = pytest.mark.django_db
 
 
 class TestReleaseHandler:
+    def test_filter_release_history(self):
+        gateway = G(Gateway)
+        stage_prod = G(Stage, gateway=gateway, name="prod")
+        stage_test = G(Stage, gateway=gateway, name="test")
+        resource_version_1 = G(ResourceVersion, gateway=gateway)
+        resource_version_2 = G(ResourceVersion, gateway=gateway)
+
+        G(ReleaseHistory, gateway=gateway, stage=stage_prod, resource_version=resource_version_1)
+        G(ReleaseHistory, gateway=gateway, stage=stage_prod, resource_version=resource_version_1, created_by="admin")
+        G(
+            ReleaseHistory,
+            gateway=gateway,
+            stage=stage_prod,
+            resource_version=resource_version_1,
+            created_time=dummy_time.time,
+        )
+        G(ReleaseHistory, gateway=gateway, stage=stage_test, resource_version=resource_version_2)
+
+        data = [
+            {
+                "params": {
+                    "query": "prod",
+                },
+                "expected": {
+                    "count": 3,
+                },
+            },
+            {
+                "params": {
+                    "stage_id": stage_prod.id,
+                },
+                "expected": {
+                    "count": 3,
+                },
+            },
+            {
+                "params": {
+                    "created_by": "adm",
+                },
+                "expected": {
+                    "count": 1,
+                },
+            },
+            {
+                "params": {
+                    "time_start": dummy_time.time - datetime.timedelta(hours=1),
+                    "time_end": dummy_time.time + datetime.timedelta(hours=1),
+                },
+                "expected": {
+                    "count": 1,
+                },
+            },
+        ]
+        for test in data:
+            result = ReleaseHandler.filter_release_history(gateway, fuzzy=True, **test["params"])
+            assert result.count() == test["expected"]["count"]
+
+    def test_filter_release_history_selects_data_plane(
+        self, django_assert_num_queries, fake_gateway, default_data_plane
+    ):
+        stage = G(Stage, gateway=fake_gateway)
+        resource_version = G(ResourceVersion, gateway=fake_gateway)
+        for _ in range(3):
+            G(
+                ReleaseHistory,
+                gateway=fake_gateway,
+                stage=stage,
+                resource_version=resource_version,
+                data_plane=default_data_plane,
+            )
+
+        queryset = ReleaseHandler.filter_release_history(fake_gateway, order_by="id")
+
+        with django_assert_num_queries(1):
+            data_plane_names = [history.data_plane.name for history in queryset]
+
+        assert data_plane_names == [default_data_plane.name] * 3
+
     def test_get_released_stage_ids(self, fake_gateway):
         fake_gateway.status = GatewayStatusEnum.ACTIVE.value
         fake_gateway.save()
